Remove whitelist terminology.
PR to change out the term whitelist in favour of allowlist.

Please note that I reference: `http://github.com/jondade/IP-allowlist-cron` instead of where the project currently resides:  http://github.com/jondade/IP-whitelist-cron in the notes. I've also Capitalised Fastly in a few places and updated `API_URL` as a static var although it doesn't appear to be in use.